Some Documentation
Describes functions of the different sofware modes for each of the subsystems. Since it'll be public and it describes the physical robot and mentions its future, I thought Mr. Eveland would appreciate it if I encrypted the word doc. Same password as the description of auton, which you can find as a pinned message on the programming slack channel.